+++ framy_generator/lib/framy_generator.dart
@@ -23,6 +23,7 @@ import 'package:framy_generator/generator/model_constructor_map_generator.dart';
 import 'package:framy_generator/generator/preset_dropdown_generator.dart';
 import 'package:framy_generator/generator/presets_map_generator.dart';
 import 'package:framy_generator/generator/routing_generator.dart';
+import 'package:framy_generator/generator/storyboard_generator.dart';
 import 'package:framy_generator/generator/subdependencies_map_generator.dart';
 import 'package:framy_generator/generator/text_field_generator.dart';
 import 'package:framy_generator/generator/toggle_page_generator.dart';
@@ -81,6 +82,7 @@ class FramyGenerator extends GeneratorForAnnotation<FramyApp> {
     buffer.writeln(generateWidgetPages(widgetFramyObjects));
     buffer.writeln(generateDependencyModel());
     buffer.writeln(generateWidgetDependenciesPanel());
+    buffer.writeln(generateStoryboardPage(widgetFramyObjects, modelFramyObjects));
     buffer.writeln(
         '\n// ======================== DEPENDENCY INPUTS ===========================\n');
     buffer.writeln(generateWidgetDependencyInput(modelFramyObjects));