+++ framy_generator/lib/framy_object.dart
@@ -131,6 +131,7 @@ enum FramyObjectType {
   preset,
   enumModel,
   riverpod,
+  page,
 }
 
 enum FramyDependencyType { constructor, provider, riverpod }