+++ framy_generator/lib/generator/drawer_generator.dart
@@ -70,6 +70,26 @@ class FramyDrawer extends StatelessWidget {
                 ),
               ),
               ${_generateCustomWidgetTiles(widgetFramyObjects)}
+              ListTile(
+                leading: Icon(Icons.view_carousel),
+                title: Wrap(
+                  crossAxisAlignment: WrapCrossAlignment.center,
+                  spacing: 8,
+                  runSpacing: 8,
+                  children: [
+                    Text('Storyboard'),
+                    Chip(
+                      label: Text(
+                        'Preview',
+                        style: TextStyle(color: Theme.of(context).primaryColor),
+                      ),
+                      backgroundColor: Colors.transparent,
+                    ),
+                  ],
+                ),
+                onTap: () =>
+                    Navigator.of(context).pushReplacementNamed('/storyboard'),
+              ),
             ],
           ),
         ),
@@ -81,23 +101,26 @@ class FramyDrawer extends StatelessWidget {
 
 String _generateCustomWidgetTiles(List<FramyObject> objects) {
   final widgetObjects = objects
-      .where((framyObject) => framyObject.type == FramyObjectType.widget)
+      .where((framyObject) =>
+  framyObject.type == FramyObjectType.widget ||
+      framyObject.type == FramyObjectType.page)
       .toList();
 
   final groupNames = widgetObjects.map((o) => o.widgetGroupName).toSet();
 
   String result = groupNames.fold(
     '',
-    (prev, groupName) => prev + _generateGroupTiles(groupName, widgetObjects),
+        (prev, groupName) =>
+    prev + _generateGroupTiles(groupName, widgetObjects),
   );
 
   result = widgetObjects
       .where((element) => element.widgetGroupName == null)
       .toList()
       .fold(
-        result,
+    result,
         (prev, framyObject) => prev + _generateTile(framyObject.name),
-      );
+  );
   return result;
 }